feat(Environment): Add Backend API Url to Angular App Environment
This commit adds the Backend API Url to the Angular App Environment in order to store it in a single place and make it easily configurable. This will allow us to switch between development and production environments without having to change the API Url in multiple places.

The implementation of this feature follows best practices for Angular development and is fully tested. This feature will improve the maintainability and scalability of our application.