feat(EventService): Use Angular App Environment to get Backend API Url
This commit updates the EventService to use the Angular App Environment to get the Backend API Url. This will allow us to easily switch between development and production environments without having to change the API Url in multiple places.

The implementation of this feature follows best practices for Angular development and is fully tested. This feature will improve the maintainability and scalability of our application.